Adding Category , Server classes, and fetching Categories from database
