+++ app/src/main/java/com/example/unishop/api/ConsultantsAPI.java
@@ -11,7 +11,6 @@ import com.android.volley.VolleyError;
 import com.android.volley.toolbox.JsonArrayRequest;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
-import com.example.unishop.R;
 import com.example.unishop.models.ModelConsultant;
 import com.example.unishop.services.ConsultantListener;
 import com.example.unishop.services.ConsultantsListener;
