+++ app/src/main/java/com/example/unishop/services/CategoriesListener.java
@@ -1,0 +1,14 @@
+package com.example.unishop.services;
+
+import com.android.volley.VolleyError;
+import com.example.unishop.models.ModelCategory;
+
+import org.json.JSONException;
+
+import java.util.List;
+
+public interface CategoriesListener {
+    void onVolleyErrorResponse(VolleyError error);
+    void onJSONObjectException(JSONException e);
+    void onCategoriesReceived(List<ModelCategory> categoryList);
+}
